getting long simulations to work

+++ notebook/combined_paper_workflow.R
@@ -18,13 +18,19 @@ source(file.path(lake_directory, "R/process_functions/glmtools.R"))
 source(file.path(lake_directory, "R/download_functions/NEON_downloads.R"))
 
 sites <- c("BARC", "CRAM", "LIRO", "PRLA", "PRPO", "SUGG")
-sites <- c("CRAM", "LIRO", "PRLA", "PRPO", "SUGG")
+
+sites <- c("CRAM")
+
+start_from_scratch <- FALSE
+time_start_index <- 36
+
 
 sim_names <- "ms_glm_flare"
 config_files <- paste0("configure_flare_",sites,".yml")
 
-num_forecasts <- 20
-days_between_forecasts <- 7
+#num_forecasts <- 20
+num_forecasts <- 19 * 7 + 1
+days_between_forecasts <- 1
 forecast_horizon <- 34 #32
 starting_date <- as_date("2021-04-18")
 second_date <- starting_date + months(1) - days(days_between_forecasts)
@@ -37,12 +43,13 @@ for(i in 3:num_forecasts){
 
 start_dates <- as_date(start_dates)
 forecast_start_dates <- start_dates + days(days_between_forecasts)
+forecast_start_dates <- forecast_start_dates[-1]
 
 configure_run_file <- "configure_run.yml"
 
 for(j in 1:length(sites)){
 
-#function(i, sites, lake_directory, sim_names, config_files, )
+  #function(i, sites, lake_directory, sim_names, config_files, )
 
   message(paste0("Running site: ", sites[j]))
 
@@ -51,13 +58,23 @@ for(j in 1:length(sites)){
   run_config$sim_name <- sim_names
   yaml::write_yaml(run_config, file = file.path(lake_directory, "configuration", "FLAREr", configure_run_file))
 
-  if(file.exists(file.path(lake_directory, "restart", sites[j], sim_names, configure_run_file))){
-    unlink(file.path(lake_directory, "restart", sites[j], sim_names, configure_run_file))
-  }
 
-  ##'
-  # Set up configurations for the data processing
-  config <- FLAREr::set_configuration(configure_run_file,lake_directory)
+
+  if(start_from_scratch){
+    FLAREr::delete_restart(sites[j], sim_names)
+    if(file.exists(file.path(lake_directory, "restart", sites[j], sim_names, configure_run_file))){
+      unlink(file.path(lake_directory, "restart", sites[j], sim_names, configure_run_file))
+    }
+    config <- FLAREr::set_configuration(configure_run_file,lake_directory)
+    config$run_config$start_datetime <- as.character(paste0(start_dates[1], " 00:00:00"))
+    config$run_config$forecast_start_datetime <- as.character(paste0(start_dates[2], " 00:00:00"))
+    config$run_config$forecast_horizon <- 0
+    config$run_config$restart_file <- NA
+    run_config <- config$run_config
+    yaml::write_yaml(run_config, file = file.path(config$file_path$configuration_directory, "FLAREr", configure_run_file))
+  }else{
+    config <- FLAREr::set_configuration(configure_run_file,lake_directory)
+  }
 
   depth_bins <- config$default_init$temp_depths
   message("    Downloading NEON data")
@@ -110,30 +127,29 @@ for(j in 1:length(sites)){
 
   cycle <- "00"
 
-  for(i in 2:length(forecast_start_dates)){
-    noaa_forecast_path <- file.path(config$met$forecast_met_model, config$location$site_id, forecast_start_dates[i], "00")
-    FLAREr::get_driver_forecast(lake_directory, forecast_path = noaa_forecast_path)
-  }
+  #for(i in time_start_index:length(forecast_start_dates)){
+  #  noaa_forecast_path <- file.path(config$met$forecast_met_model, config$location$site_id, forecast_start_dates[i], "00")
+  #  FLAREr::get_driver_forecast(lake_directory, forecast_path = noaa_forecast_path)
+  #}
 
-  FLAREr::get_stacked_noaa(lake_directory, config, averaged = TRUE)
+  #FLAREr::get_stacked_noaa(lake_directory, config, averaged = TRUE)
 
-  config$run_config$start_datetime <- as.character(paste0(start_dates[1], " 00:00:00"))
-  config$run_config$forecast_start_datetime <- as.character(paste0(start_dates[2], " 00:00:00"))
-  config$run_config$forecast_horizon <- 0
-  config$run_config$restart_file <- NA
-  run_config <- config$run_config
-  yaml::write_yaml(run_config, file = file.path(config$file_path$configuration_directory, "FLAREr", configure_run_file))
+  for(i in time_start_index:length(forecast_start_dates)){
 
-  for(i in 1:length(forecast_start_dates)){
+    unlist(file.path(lake_directory, "flare_tempdir"), recursive = TRUE)
 
     config <- FLAREr::set_configuration(configure_run_file,lake_directory)
     config <- FLAREr::get_restart_file(config, lake_directory)
 
-    message(paste0("     Running forecast that starts on: ", config$run_config$start_datetime))
+    message(paste0("     Running forecast that starts on: ", config$run_config$start_datetime, " with index: ",i))
 
+    config$file_path$noaa_directory <- "/data/drivers"
     if(config$run_config$forecast_horizon > 0){
-      noaa_forecast_path <- FLAREr::get_driver_forecast_path(config,
-                                                             forecast_model = config$met$forecast_met_model)
+
+      noaa_forecast_path <- file.path(config$met$forecast_met_model, config$location$site_id, forecast_start_dates[i], "00")
+      #noaa_forecast_path <- FLAREr::get_driver_forecast_path(config,
+      #                                                       forecast_model = config$met$forecast_met_model)
+      #FLAREr::get_driver_forecast(lake_directory, forecast_path = noaa_forecast_path)
       forecast_dir <- file.path(config$file_path$noaa_directory, noaa_forecast_path)
     }else{
       forecast_dir <- NULL
@@ -170,6 +186,8 @@ for(j in 1:length(sites)){
                                                 restart_file = config$run_config$restart_file,
                                                 historical_met_error = met_out$historical_met_error)
 
+    unlink(config$run_config$restart_file)
+
     ##' Run the forecasts
     message("    Starting Data Assimilation and Forecasting")
     da_forecast_output <- FLAREr::run_da_forecast(states_init = init$states,
@@ -199,14 +217,25 @@ for(j in 1:length(sites)){
     rm(da_forecast_output)
     gc()
     message("Generating plot")
-    FLAREr::plotting_general_2(file_name = saved_file,
-                               target_file = file.path(config$file_path$qaqc_data_directory, paste0(config$location$site_id, "-targets-insitu.csv")),
-                               ncore = 2,
-                               obs_csv = FALSE)
+    pdf_file <- FLAREr::plotting_general_2(file_name = saved_file,
+                                           target_file = file.path(config$file_path$qaqc_data_directory, paste0(config$location$site_id, "-targets-insitu.csv")),
+                                           ncore = 2,
+                                           obs_csv = FALSE)
+
+    FLAREr::put_forecast(saved_file, eml_file_name, config)
+
+    if(config$run_config$use_s3){
+      success <- aws.s3::put_object(file = pdf_file, object = file.path(config$location$site_id, basename(pdf_file)), bucket = "analysis")
+      if(success){
+        unlink(pdf_file)
+      }
+    }
 
     message(paste0("Metadata and plots generated! Go to NEON-forecast/forecasts/",config$location$site_id," folder to view"))
 
     FLAREr::update_run_config(config, lake_directory, configure_run_file, saved_file, new_horizon = forecast_horizon, day_advance = days_between_forecasts)
+
+    unlist(forecast_dir, recursive = TRUE)
 
     if(i > 1){
 
@@ -256,7 +285,16 @@ for(j in 1:length(sites)){
 
       forecast_file <- paste(sites[j], as_date(config$run_config$forecast_start_datetime), "ms_climatology.csv.gz", sep = "-")
 
-      write_csv(clim_forecast, file = file.path(config$file_path$forecast_output_directory, forecast_file))
+      saved_file <- file.path(config$file_path$forecast_output_directory, forecast_file)
+      write_csv(clim_forecast, file = saved_file)
+
+      if (config$run_config$use_s3) {
+        success <- aws.s3::put_object(file = saved_file, object = file.path(config$location$site_id,
+                                                                            basename(saved_file)), bucket = "forecasts")
+        if (success) {
+          unlink(saved_file)
+        }
+      }
 
       message("   creating persistence forecast")
 
@@ -298,8 +336,16 @@ for(j in 1:length(sites)){
       }
 
       forecast_file <- paste(sites[j], as_date(config$run_config$forecast_start_datetime), "ms_persistence.csv.gz", sep = "-")
-
+      saved_file <- file.path(config$file_path$forecast_output_directory, forecast_file)
       write_csv(persist_forecast, file = file.path(config$file_path$forecast_output_directory, forecast_file))
+
+      if (config$run_config$use_s3) {
+        success <- aws.s3::put_object(file = saved_file, object = file.path(config$location$site_id,
+                                                                            basename(saved_file)), bucket = "forecasts")
+        if (success) {
+          unlink(saved_file)
+        }
+      }
     }
   }
 }
